Allow for data and summaries in the payload.
