+++ pkg/sqlcache/informer/db_mocks_test.go
@@ -206,19 +206,19 @@ func (mr *MockClientMockRecorder) ReadObjects(rows, typ any) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReadObjects", reflect.TypeOf((*MockClient)(nil).ReadObjects), rows, typ)
 }
 
-// ReadStringIntString mocks base method.
-func (m *MockClient) ReadStringIntString(rows db.Rows) ([][]string, error) {
+// ReadStringIntString1or2 mocks base method.
+func (m *MockClient) ReadStringIntString1or2(rows db.Rows, readThirdString bool) ([][]string, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ReadStringIntString", rows)
+	ret := m.ctrl.Call(m, "ReadStringIntString1or2", rows, readThirdString)
 	ret0, _ := ret[0].([][]string)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
-// ReadStringIntString indicates an expected call of ReadStringIntString.
-func (mr *MockClientMockRecorder) ReadStringIntString(rows any) *gomock.Call {
+// ReadStringIntString1or2 indicates an expected call of ReadStringIntString1or2.
+func (mr *MockClientMockRecorder) ReadStringIntString1or2(rows, readThirdString any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReadStringIntString", reflect.TypeOf((*MockClient)(nil).ReadStringIntString), rows)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReadStringIntString1or2", reflect.TypeOf((*MockClient)(nil).ReadStringIntString1or2), rows, readThirdString)
 }
 
 // ReadStrings mocks base method.
